Consolidate watch invalidation tests into existing suites
Move related watch cases into existing compiler-watch, snapshot, and
native watcher suites while preserving package and platform boundaries.

Keep every test, benchmark, helper, assertion, and fixture. Update the
regular-file pathconf fixture to refer to its surviving test file.
